rework stream cli cmds
