+++ pkg/streams/cli/dev_set.go
@@ -17,9 +17,10 @@ func newDeviceSetCmd() *cobra.Command {
 	bucket := config.DeviceBucket
 
 	cmd := &cobra.Command{
-		Use:     "set",
+		Use:     "device-set",
 		Short:   "Create or update a device profile",
 		Aliases: []string{"update"},
+		GroupID: "device",
 		RunE: func(cmd *cobra.Command, _ []string) error {
 			return withDeviceStore(cmd.Context(), bucket, func(mgr *store.DeviceStore) error {
 				if bufferDur > 0 {
@@ -36,13 +37,13 @@ func newDeviceSetCmd() *cobra.Command {
 		},
 	}
 
-	cmd.Flags().StringVar(&deviceID, "device-id", "", "Device identifier")
+	cmd.Flags().StringVar(&deviceID, "device", "", "Device identifier")
 	cmd.Flags().StringVar(&bucket, "device-bucket", bucket, "Device metadata bucket")
 	cmd.Flags().StringVar(&info.Description, "description", "", "Device description")
 	cmd.Flags().StringVar(&info.Location, "location", "", "Device location")
 	cmd.Flags().StringVar(&info.Owner, "owner", "", "Device owner")
 	cmd.Flags().DurationVar(&bufferDur, "buffer", 0, "Optional rolling buffer window (e.g. 5m)")
-	if err := cmd.MarkFlagRequired("device-id"); err != nil {
+	if err := cmd.MarkFlagRequired("device"); err != nil {
 		cobra.CheckErr(err)
 	}
