+++ pkg/streams/cli/rec_play.go
@@ -11,7 +11,7 @@ import (
 	"github.com/spf13/cobra"
 )
 
-func newRecordingsPlayCmd() *cobra.Command {
+func newStreamPlayCmd() *cobra.Command {
 	opts := &session.PlaybackOptions{
 		Bucket: config.SessionBucket,
 		Speed:  1,
@@ -19,8 +19,9 @@ func newRecordingsPlayCmd() *cobra.Command {
 
 	cmd := &cobra.Command{
 		Use:     "play",
-		Short:   "Replay a recorded session",
+		Short:   "play back a recorded stream session",
 		Aliases: []string{"replay"},
+		GroupID: "record",
 		RunE: func(cmd *cobra.Command, _ []string) error {
 			ctx, cancel := signal.NotifyContext(context.Background(), os.Interrupt, syscall.SIGTERM)
 			defer cancel()
@@ -31,11 +32,11 @@ func newRecordingsPlayCmd() *cobra.Command {
 		},
 	}
 
-	cmd.Flags().StringVar(&opts.SessionID, "session-id", "", "Session identifier to replay")
+	cmd.Flags().StringVar(&opts.SessionID, "session", "", "Session identifier to replay")
 	cmd.Flags().StringVar(&opts.TargetSubject, "target-subject", "", "Optional override subject to publish during playback (default: "+config.PlaybackSubject+")")
 	cmd.Flags().Float64Var(&opts.Speed, "speed", opts.Speed, "Playback speed multiplier (e.g. 0.25, 1, 5)")
 	cmd.Flags().StringVar(&opts.Bucket, "session-bucket", opts.Bucket, "Key-value bucket containing session metadata")
-	if err := cmd.MarkFlagRequired("session-id"); err != nil {
+	if err := cmd.MarkFlagRequired("session"); err != nil {
 		cobra.CheckErr(err)
 	}
